Добавлены исправления по замечаниям ревьюера
