+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -3,9 +3,11 @@ package ru.yandex.practicum.filmorate.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Slf4j
@@ -30,12 +32,14 @@ public class UserController {
     @PostMapping
     public User createUser(@RequestBody User user) {
         log.debug("POST /users with {}", user);
+        validate(user);
         return userService.createUser(user);
     }
 
     @PutMapping
     public User updateUser(@RequestBody User user) {
         log.debug("PUT /users with {}", user);
+        validate(user);
         return userService.updateUser(user);
     }
 
@@ -61,5 +65,29 @@ public class UserController {
     public List<User> getCommonFriends(@PathVariable Integer id, @PathVariable Integer otherId) {
         log.debug("GET /users/{id}/friends/common by id {} and otherId {}", id, otherId);
         return userService.getCommonFriends(id, otherId);
+    }
+
+    void validate(User user) {
+        log.debug("Validation start for {}", user);
+
+        if (user.getEmail() == null || user.getEmail().isBlank() || !user.getEmail().contains("@")) {
+            String message = "Email пользователя не может быть пустым и должен содержать символ @! email = "
+                    + user.getEmail();
+            log.error(message);
+            throw new ValidationException(message);
+        }
+        if (user.getLogin() == null || user.getLogin().isBlank() || user.getLogin().contains(" ")) {
+            String message = "Логин не может быть пустым и содержать пробелы!";
+            log.error(message);
+            throw new ValidationException(message);
+        }
+        if (user.getBirthday().isAfter(LocalDate.now())) {
+            String message = "Дата рождения не может быть в будущем!";
+            log.error(message);
+            throw new ValidationException(message);
+        }
+        if (user.getName() == null || user.getName().isBlank()) {
+            user.setName(user.getLogin());
+        }
     }
 }
