+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -2,11 +2,8 @@ package ru.yandex.practicum.filmorate.storage.film;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -22,8 +19,9 @@ public class InMemoryFilmStorage implements FilmStorage {
     public Film getFilmById(Integer id) {
         if (filmStorage.containsKey(id)) {
             return filmStorage.get(id);
+        } else {
+            return null;
         }
-        throw new NotFoundException("Такого фильма нет!");
     }
 
     @Override
@@ -34,7 +32,6 @@ public class InMemoryFilmStorage implements FilmStorage {
     @Override
     public Film createFilm(Film film) {
         log.debug("POST /films with {}", film);
-        validate(film);
         film.setId(getNextId());
         filmStorage.put(film.getId(), film);
         return film;
@@ -44,9 +41,8 @@ public class InMemoryFilmStorage implements FilmStorage {
     public Film updateFilm(Film film) {
         log.debug("PUT /films with {}", film);
         if (!filmStorage.containsKey(film.getId())) {
-            throw new NotFoundException("Такого фильма нет, обновление невозможно!");
+            return null;
         }
-        validate(film);
         Film filmFromStorage = filmStorage.get(film.getId());
         filmFromStorage.setId(film.getId());
         filmFromStorage.setName(film.getName());
@@ -55,33 +51,6 @@ public class InMemoryFilmStorage implements FilmStorage {
         filmFromStorage.setReleaseDate(film.getReleaseDate());
         filmStorage.put(film.getId(), filmFromStorage);
         return filmFromStorage;
-    }
-
-    void validate(Film film) {
-        log.debug("Validation start for {}", film);
-
-        final LocalDate DATE_OF_FIRST_FILM = LocalDate.of(1895, 12, 28);
-
-        if (film.getName() == null || film.getName().isEmpty()) {
-            String message = "Имя фильма не может быть пустым! name = " + film.getName();
-            log.error(message);
-            throw new ValidationException(message);
-        }
-        if (film.getDescription().length() > 200) {
-            String message = "Максимальная длина описания - 200 символов! length = " + film.getDescription().length();
-            log.error(message);
-            throw new ValidationException(message);
-        }
-        if (film.getReleaseDate().isBefore(DATE_OF_FIRST_FILM)) {
-            String message = "Дата релиза не может быть раньше 28 декабря 1895 года! date = " + film.getReleaseDate();
-            log.error(message);
-            throw new ValidationException(message);
-        }
-        if (film.getDuration() < 0) {
-            String message = "Продолжительность фильма не может быть меньше нуля! duration = " + film.getDuration();
-            log.error(message);
-            throw new ValidationException(message);
-        }
     }
 
     private int getNextId() {
